dom/tests/mochitest/chrome/DOMWindowCreated_chrome.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>
<!--
  Test for Persistent Storage in chrome
  -->
<window id="sample-window" width="400" height="400"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" 
          src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript" 
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>      

<script type="application/javascript">
var SimpleTest = window.opener.wrappedJSObject.SimpleTest;

document.addEventListener("DOMWindowCreated", function(e) {
  var otherWindow = e.target.defaultView.wrappedJSObject;
  SimpleTest.is(e.type, "DOMWindowCreated", "DOMWindowCreated: " + otherWindow);
  otherWindow.doneFunction = function() {
    SimpleTest.ok(true, "doneFunction was called");
    SimpleTest.finish();
    window.close();
  };
}, false);
</script>

<iframe type="content" src="http://mochi.test:8888/chrome/dom/tests/mochitest/chrome/DOMWindowCreated_content.html"/>

</window>
